UBUNTU: snd_bt_sco kernel Oops at startup OriginalAuthor: Colin Ian King Bug: #202249
btsco.c requires the correct ALSA build flags to make sure
struct snd_pcm_str in pcm.h is defined the same way as the ALSA
core to avoid the kernel Oops. We need to add in the appropriate
flags as currently btsco.o is not part of the ALSA tree.

Signed-off-by: Colin Ian King <colin.king@canonical.com>