kgdb: Implement kgdb over ethernet with NET_POLL
At one point this was very similar to the version Matt Mackall wrote
to allow for the use of KGDB over ethernet.  Since then it has been
reworked to fit into the unified KGDB model.

This kgdb I/O driver can work as a kernel built-in or a kernel
module.

Signed-off-by: Tom Rini <trini@kernel.crashing.org>
Signed-off-by: Jason Wessel <jason.wessel@windriver.com>
Acked-by: Matt Mackall <mpm@selenic.com>