kgdb, mips: This adds basic support to the MIPS architecture
This patch removes the gdb-stub for mips and replaces it with an
architecture specific implementation to work with the kgdb core.  This
required the removal of a number of board specific gdb-stub hacks.

The new kgdb architecture specific handler registers and unregisters
dynamically for exceptions depending on when you configure a kgdb I/O
driver.

There have been quite a number of contributors during the existence
of this patch (see arch/mips/kernel/kgdb.c).

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>