UBUNTU: Fix potential lockups locking OriginalAuthor: Daniel Mack <daniel@caiaq.org> OriginalLocation: http://article.gmane.org/gmane.linux.alsa.devel/52464
attached is a patch for snd_usb_caiaq which fixes system potential
lockups when stoping and restarting the stream very quickly multiple
times in a row which is what some players do when the user drags the
song position slider.

Also, it seems that snd_pcm_period_elapsed() must not be called when
holding a spinlock? The documentation[*], however, doesn't state that.

As this fixes an annoying bug, I would appreciate seeing that one being
pushed into 2.6.25. Still possible?

Thanks and best regards,
Daniel

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>