UBUNTU: Remove openvm tools OriginalAuthor: Soren Hansen soren@ubuntu.com
We're reliably informed that open-vm-tools is not something we
want to support. Each VMWare product comes with its own version of
these kernel modules (released under the GPL and everything),
while open-vm-tools is not actually supported for any version of
any VMWare product. This might very well change in either the
intrepid or intrepid+1 time frame, but it's not there yet.

Signed-off-by: Tim Gardner <tim.gardner@canonical.com>