UBUNTU: Use LUM version of ALSA headers when building media drivers.
The ALSA version in LUM is used to replace the kernel code. All code that
want's to use the ALSA driver must use the new headers. Some drivers have
been moved to LUM for this reason but did not use the right headers.
This is now changed for all drivers in the media directory. Unfortunately
this could not be done for the complete build of LUM since there are con-
flicting files for some drivers.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>