UBUNTU: build: Include correct ALSA config for media drivers. Bug: #212960
When building drivers that depend on ALSA it is not only important to have
the correct headers but also to have the right configuration settings
included, since the ALSA headers change based on those values.
When building from within the kernel this is done automatically by the
kernel build environment. But in our case we have to include the settings
from ALSA.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>