UBUNTU: Allow unionfs to chown a hardlinked file on a ro filesystem OriginalAuthor: Colin Ian King Bug: #224754
This fixes a problem with chown'ing a file which has the suid bit set
and has more than one hardlink on a read-only file system. The stacked
nature of the unionfs means that the attribute changes pass through the
lower level attr fs routines with the ATTR_MODE flag set when it's already
been handled before in a stacked operation, so clear the ATTR_MODE flag.
(This is the same fix as in the stacked ecrypt fs.)

Signed-off-by: Colin Ian King <colin.king@canonical.com>