UBUNTU: Enable appleir driver for new MacBook Pro OriginalAuthor: Alex Karpenko OriginalLocation: http://launchpadlibrarian.net/16351456/appleir_add_device_id_for_new_macbooks.patch Bug: #157919
Patch from Alex Karpenko to enable appleir for new MacBook Pro with
IR device ID of 0x8242. In addition, the patch changes the key codes
for back and forward to correspond with those used on apple keyboards.
That way media controls on the remote work out of the box
(e.g., in rhythmbox)

Signed-off-by: Colin Ian King <colin.king@canonical.com>