mainline builds: primative first stab at a cronable trigger script
We want to be able to automate triggering of these builds.  This is the
first stab at a script to handle triggering new builds on a completely
automated basis.

NOTE: this does require the build directory to already exist and for
the tree to contain the requisite remotes, it does not build the tree.

Signed-off-by: Andy Whitcroft <apw@canonical.com>