UBUNTU: dib0700: prevent previous changeset from affecting older devices
OriginalAuthor: Michael Krufky <mkrufky@linuxtv.org>

The previous changeset updates the DIB0700_DEFAULT_DEVICE_PROPERTIES
to use the 1.20 firmware for all devices.  We might not want to make
this drastic change in a stable kernel, so better off to only use the
new firmware only where it is actually necessary.

This changeset reverts DIB0700_DEFAULT_DEVICE_PROPERTIES to use the
1.10 firmware, and creates DIB0700_FW_1_20_DEVICE_PROPERTIES for the
1.20 firmware.

Signed-off-by: Michael Krufky <mkrufky@linuxtv.org>