UBUNTU: add option, "disable_streaming_master_mode" OriginalAuthor: Devin Heitmueller <devin.heitmueller@gmail.com> OriginalLocation: http://linuxtv.org/hg/v4l-dvb/rev/a79b809b763a
This mode of the dib0700 is required for usage of any demodulator
other than the dib0x00 demods.

This was backported from an original patch that adds support for
the PCTV HD Pro 801e. The bits that add the new device support
have been held back due to missing driver dependencies, such as
xc5000 and s5h1411.

Signed-off-by: Devin Heitmueller <devin.heitmueller@gmail.com>
Signed-off-by: Patrick Boettcher <pb@linuxtv.org>
Signed-off-by: Michael Krufky <mkrufky@linuxtv.org>