UBUNTU: SAUCE: reorder acpi/Makefile avoiding sysfs duplicate insert warning
BugLink: https://bugs.launchpad.net/ubuntu/+source/linux/+bug/355977

We are seeing the following kernel warning on Hardy netbook kernels:

    [ 9.459204] WARNING: at .../fs/sysfs/dir.c:429 sysfs_add_one()
    [ 9.459210] Pid: 1, comm: swapper Not tainted 2.6.24-24-lpia #3
    [ 9.459213] [<c01d2754>] sysfs_add_one+0xd4/0x110
    [ 9.459220] [<c01d2c88>] create_dir+0x48/0x90
    [ 9.459227] [<c01d2cf9>] sysfs_create_dir+0x29/0x50
    [ 9.459233] [<c023cb3f>] kobject_get+0xf/0x20
    [ 9.459239] [<c023d003>] kobject_add+0x93/0x1b0
    [ 9.459245] [<c023cf41>] kobject_set_name+0x81/0xb0
    [ 9.459252] [<c0551c44>] kernel_param_sysfs_setup+0x54/0xa0
    [ 9.459259] [<c0551d9d>] param_sysfs_init+0x10d/0x190
    [ 9.459266] [<c053e551>] kernel_init+0x131/0x320
    [ 9.459271] [<c010435e>] __switch_to+0x9e/0x150
    [ 9.459278] [<c01052c6>] ret_from_fork+0x6/0x20
    [ 9.459284] [<c053e420>] kernel_init+0x0/0x320
    [ 9.459289] [<c053e420>] kernel_init+0x0/0x320
    [ 9.459295] [<c0106667>] kernel_thread_helper+0x7/0x10
    [...]
    [ 9.459453] Module 'acpi' failed to be added to sysfs, error number -17

We are getting this message because we have two modules (both builtin)
which are using acpi. prefixes for their modules parameters.  This triggers
the code in param_sysfs_builtin() to find two acpi modules and attempt
to create the acpi module parameters directory in sysfs twice, tripping
the warning.

Reorder the linking of the acpi modules to place these together causing
their module entries to appear consecutivly in the module parameter
table and prevent the warning.

Signed-off-by: Andy Whitcroft <apw@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Colin King <colin.king@canonical.com>