Revert "UBUNTU: Bluetooth: SCO flow control to enable bluetooth headsets OriginalLocation: http://bluetooth-alsa.cvs.sourceforge.net/*checkout*/bluetooth-alsa/plugz/patches/sco-flowcontrol-v4.4.diff"
These are against ubuntu-hardy's netbook-lpia branch.  They probably
shouldn't be committed until we Ulisses's userspace patches are ready for
netbook-common, but it would be good to get the kernel team's ACKs for now.

This reverts commit 20424eae73ff0bc78802b8fdf5b2bc9c0d587772.

As Marcel points out in #130870, this patch is incorrect.  The bug is
fixed in later bluez userspace packages.

Signed-off-by: Andres Salomon <dilinger@canonical.com>