UBUNTU: SAUCE: Enable debug package generation
BugLink: https://bugs.launchpad.net/ubuntu/+source/linux/+bug/340181

Enable debug package generation as part of the standard build.
This works for both main builds and custom builds.  The original
method did not follow the build systems convention and the custom
build did not have debug build handling at all.  Debug builds now
follow the same method as Intrepid and Jaunty builds.

The default case is not to create debug (.ddeb) packages.  This applies
to local builds.

A debug package will be created in the build environment (chroot or VM)
if and only if the directory '/CurrentlyBuilding' exists.  To turn on
a debug build, all that is necessary is:

  sudo mkdir /CurrentlyBuilding

before the build.  This is the method used on the build systems.

There were some environment variables that were set per architecture.  These
did not consistantly work and were removed.  The build systems do not use
these special environment variables.  They will not be missed.

Signed-off-by: Jim Lieb <lieb@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>