allow hosts to fetch git trees from the DC to seed pushes
When pushing a git tree for the first time the whole tree history must
be pushed.  In a lot of cases it is far quicker to download the majority
of the tree from the DC copy.  This is especially true for porters which
are on in the same location.  Add a new per host variable USE_DC=yes
which enables pre-fetching the DC copies of the trees before pushing
your changes.  The git repo is determined from the repo connected
to the primary build repo's origin.

Signed-off-by: Andy Whitcroft <apw@canonical.com>