UBUNTU: printchanges -- rebase tree does not have stable tags use changelog
When generating the changes list we normally use the previous tag to locate
the start of the new changes.  However with a rebase tree the tag is off
on another branch.  Instead simply search back in the log looking for the
commit which represents the previous release commit and use that directly.

Signed-off-by: Andy Whitcroft <apw@canonical.com>