UBUNTU: ALSA: snd-hda-intel: fix ALC662 capture device init
This fixes a bug for the ALC622 audio that was introduced in
commit 19a99f198f5c45119b4ebaae2b8f106dad10ea59 when fixing a bug
for the ALC272_DELL_ZM1.

Instead of adding the ALC662 capture mixer by explicitly
calling add_mixer we should be actually specifying the mixer
in the alc662_presets array for the ALC272_DELL_ZM1.

Without this fix, the driver tries to add two capture mixers
for the ALC662 which causes the driver to fail to initialise.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
(cherry picked from commit ddd6b47f04f88d4e0d9b18da486fbc6ec51419e2)

Signed-off-by: Steve Conklin <sconklin@canonical.com>