UBUNTU: SAUCE: Fix the MIC of the Lenovo webcam problem
BugLink: https://bugs.edge.launchpad.net/ubuntu/+source/linux/+bug/310760

The MIC of the Lenovo webcam doesn't work after launching the
camera, because of the i2c command that turn on/off the LED for
the camera is wrong and will impact the MIC function.

By snooping the USB traffic from Windows and extracting the correct
i2c commands that turn on/off the LED.
The patch is already accepted by the V4L/DVB branch and will be merged
into upstream kernel later.

http://linuxtv.org/hg/v4l-dvb/rev/49966c5f2052

Signed-off-by: AceLan Kao <acelan.kao@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Amit Kucheria <amit.kucheria@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>