SYSLINK: ipc - removed gt_traces in remaining IPC modules
This patch cleans up the gt_trace code in the rest of the
IPC modules. printks are added instead in the IPC driver
code. The gt.h header file is no longer included in any
IPC source files. This patch has also fixed couple of
checkpatch warnings in sharedregion module.

Signed-off-by: Suman Anna <s-anna@ti.com>