USB: android gadget: mass storage: Fix format issue for Vista Host
BugLink: http://bugs.launchpad.net/bugs/666234

As part of formating the mass storage device, Host sends the INQUIRY
scsi command. As per the standard, the command length for this
command should be 6 bytes, whereas the Vista host sends 12 bytes.

When the command length of the command is not equal to the standard
length, the device sends a phase error as part of the status phase.
When the host receives a phase error, it re-enumerates, hence the
error.

If the command is INQUIRY, and the command length is 12 bytes,
treating this as a good command and not sending the phase error
to the host fixes this issue.

Signed-off-by: Velempati Chiranjeevi <c_cvelem@qualcomm.com>
Signed-off-by: Mike Lockwood <lockwood@android.com>
(cherry picked from commit fd515513a38b53d2b961cbbc9c4fdaeedec74d44)
Signed-off-by: Saeed Bishara <saeed@marvell.com>
Signed-off-by: Eric Miao <eric.miao@canonical.com>

Signed-off-by: Tim Gardner <tim.gardner@canonical.com>