UBUNTU: [Config]: Turn on CONFIG_ARM_THUMBEE
As Dave Martin suggested, some ThumbEE based JIT code is to be merged into
openjdk-6 for lucid.

The ThumbEE extension has a small amount of additional per-process context
which will not be switched properly unless CONFIG_ARM_THUMBEE is enabled,
leading to potential instability of user processes which use this extension.

Signed-off-by: Eric Miao <eric.miao@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>