UBUNTU: [Config] Enable CONFIG_THUMBEE for fsl-imx51 branch
As Dave Martin suggested, some ThumbEE based JIT code is to be merged into
openjdk-6 for lucid.

The ThumbEE extension has a small amount of additional per-process context
which will not be switched properly unless CONFIG_ARM_THUMBEE is enabled,
leading to potential instability of user processes which use this extension.

Signed-off-by: Bryan Wu <bryan.wu@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>