ALSA: hda-intel: Avoid divide by zero crash
CVE-2010-1085

On my AMD780V chipset, hda_intel.c can crash the kernel with a divide by
zero
for as-yet unknown reasons. A simple check for zero prevents it, though
the problem that causes it remains. Since the workaround is harmless and
won't affect anyone except victims of this bug, it should be safe;
moreover,
because this crash can be triggered by a user-mode application, there are
denial of service implications on the systems affected by the bug without
the patch.

Signed-off-by: Jody Bruchon <jody@nctritech.com>
Cc: <stable@kernel.org>
Signed-off-by: Takashi Iwai <tiwai@suse.de>
(cherry-picked from commit fed08d036f2aabd8d0c684439de37f8ebec2bbc2 upstream)
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>