UBUNTU: [Config] Enable all CGROUP configuration options
BugLink: http://bugs.launchpad.net/bugs/480739

These options are needed for Linux Containers to work. We changed those
to the same values in Lucid (without finding regressions).
So now take those changes back to Karmic and also make the options for
ports match up with the ones for common.ubuntu.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Leann Ogasawara <leann.ogasawara@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>