SYSLINK: ipc - sharedregion module migrated to SysLink 2.0
SharedRegion module is updated for SysLink-2.0. This version
incorporates changes for improved ease of programming for the
user.

The new SharedRegion module is a drastically different module
compared to before. It supports heap creation in-built into
the regions provided by the new HeapMemMP module. This is used
for automatic gate creation and other shared object creation.

The following are the main changes specific to this module:
1. Old API is deprecated. It retains only a couple of
   old API.
2. A whole set of new API are provided. The list includes
   sharedregion_attach, sharedregion_detach, sharedtegion_start,
   sharedregion_stop, sharedregion_reserve_memory, sharedregion_set_entry,
   sharedregion_clear_entry and sharedregion_clear_reserved_memory
   among others.

The following are the changes in line to the common changes to all SysLink-2.0
IPC modules.
1. A sharedregion_module variable is introduced to mimic the BIOS behavior.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Angela Stegmaier <angelabaker@ti.com>