UBUNTU: Add in staging rt28x0 drivers from 2.6.34
The rt28x0sta.ko modules that we ship in Ubuntu 10.04 LTS will not
reconnect after the first disconnection for some users. The version of
the drivers from 2.6.34 work correctly for some. This commit backports
the entire rt28x0 stack from 2.6.34 and adds the modules to the
linux-backports-modules-wireless package.

I also added the new rt2870.bin firmware because it is essential for the
rt2870sta module to function. The rt2860.bin firmware has not changed
since the 10.04 LTS release.

BugLink: http://bugs.launchpad.net/bugs/496093

Signed-off-by: Chase Douglas <chase.douglas@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>