UBUNTU: SAUCE: Add thinkpad-acpi to modules in the ALSA subpackage
BugLink: http://bugs.launchpad.net/bugs/599527

The thinkpad-acpi driver links against ALSA to provide a mixer interface
for the hardware mixer of ThinkPads. Without this module being rebuild
against the LBM ALSA driver it will not load.

This patch takes the version of the thinkpad-acpi driver from Lucid (which
is a 2.6.34 version anyway) and modifies the build system to build it with
the updated symbol versions from the LBM ALSA drivers.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>