cvescripts: cve-setup: Create local tracker branch
Instead of relying on the bzr branch being created correctly and needing
a reference to it, create it locally as a part of the initial setup.
Also adds a new library which contains commands for handling the branch
(make sure to merge updates from the origin and push them to the team
branch).

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>