cvescripts: cvetracker_lib: Account for bizare strangeness
When there are changes in the master branch and unmerged local changes,
then one needs to manually commit that merge with its own message.
When those changes are merged back into the master branch there will be
a merge, which looks like replacing the history, leaving nothing to
commit but requires a push out to the team branch.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>