maintscripts: maint-getabis: Fix running on topic branches
There were two issues: first the flavour-control.stub (at least for
mvl-dove) had the "Provides:" line commented out. As this is used to
detect which package is the linux kernel, the whole process failed.
Added a better error message for that case while fixing it to work with
commented-out lines.
Second, depending on the source package name (which is different for
topic branches) the binary package is in a different subdirectory in the
archive.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>