ecryptfs: Bugfix for error related to ecryptfs_hash_buckets
CVE-2010-2492

The function ecryptfs_uid_hash wrongly assumes that the
second parameter to hash_long() is the number of hash
buckets instead of the number of hash bits.
This patch fixes that and renames the variable
ecryptfs_hash_buckets to ecryptfs_hash_bits to make it
clearer.

Signed-off-by: Andre Osterhues <aosterhues@escrypt.com>
Signed-off-by: Tyler Hicks <tyhicks@linux.vnet.ibm.com>
(picked from security list)
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>