OMAP4: powerdomain: Move powerdomain.c from mach-omap2 to plat-omap
This is in preparation of splitting the powerdomain framework into
platform-independent part (for all omaps) and platform-specific
parts.
The platform-independent code would reside in plat-omap/powerdomain.c
and the platform-specific code will resides in mach-omap2/powerdomain-xxxx.c
files.

Some of the hacks to include header files are done to make sure the patch compiles
and works fine. These hacks will be removed in the subsequent patches.

Signed-off-by; Rajendra Nayak  <rnayak@ti.com>