ARM: OMAP4: Make TWL6040 codec an MFD device
In order to have audio and vibrator driver, the twl6040 has been
turned into an MFD device. This twl6040_codec will be registered
as a child of TWL-CORE, and will have two childs of its own:

- twl6040_audio
- twl6040 vibra

This is based on Peter Ujfalusi's patch to create a twl4030 codec
MFD subdevice.

This patch also adds the twl6040_audio part.

Signed-off-by: Jorge Eduardo Candelaria <jorge.candelaria@ti.com>