maintscripts: maint-startnewrelease: Fix test for open release
There was a check in the script for when the version state in the changelog
is "UNRELEASED". However the test was failing because of an extra character
in the string in the test.

I also added some additional text to the message printed to make
it more clear that a "start new release" is not necessary.

Signed-off-by: Brad Figg <brad.figg@canonical.com>