UBUNTU: SAUCE: hid: 3m: Simplify touchsreen emulation logic
The ten-finger 3M screen M2256PW has a kernel driver already
(hid-3m-pct), but with a couple of glitches:

1. The one-touch down/up, which emits BTN_TOUCH events, sometimes does
not work properly. This results in buttons getting stuck in the
button-down state in the window manager.

2. The action after touch-up in one spot followed by a touch-down in
another spot sometimes causes a sequence of strange finger-tracing
events along the perimeter of the rectangle between the two spots. The
root of the problem lies in how the kernel inteprets the HID hybrid
event type which is used by the 3M device [1].

While a proper HID solution is being worked out upstream, this patch
simplifies the driver logic such that BTN_TOUCH and ABS_X/Y events are
emitted properly, in effect papering over the worst problems. Suggested
for inclusion in Maverick.

[1] Suggested by Stephane Chatty, who also confirms the observed problems.

Signed-off-by: Henrik Rydberg <rydberg@euromail.se>
Acked-by: Chase Douglas <chase.douglas@canonical.com>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>