SYSLINK:IPU-PM: avoiding down_timeout when failing
In ipu_pm_notifications the PID_DEATH/SUSPEND/RESUME
event is sent to IPU and a semaphore is used to wait
for IPU to respond. In the case when the send_event
is failing, there is no need to called the semaphore
and wait.
Also if one of the proc (SYS/APP) is not responding
an error is returned without sending the notification
to the other proc.

Signed-off-by: Miguel Vadillo <vadillo@ti.com>
Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>
Signed-off-by: Angela Stegmaier <angelabaker@ti.com>
Signed-off-by: Paul Hunt <hunt@ti.com>