OMAP4:DSS: add generic notifier mechanism
A callback can be registered by the dssdev client in order to be
notified of resolution changes, for example an external monitor
that is hot-plugged.

Multiple clients can now register for notification from one
dssdev, and the notification mechanism can be extended in the
future to add other events.

Note that this patch supersedes the "OMAP4:DSS: add callback to notify
client of resolution change" patch.

Signed-off-by: Rob Clark <rob@ti.com>