OMAP4: PandaBoard: add board revision reporting
Add initial support for determining the board revision of the
PandaBoard and reports the version on boot. This will assist in
determining which board is being used when providing techincal
support. The board revision will also be used to determine which
GPIO and other configuration changes are needed between different
hardware revisions of the PandaBoard.

Signed-off-by: David Anders <x0132446@ti.com>