UBUNTU: bodge linux-libc-dev package version due to ti-omap4 error
The ti-omap4 2.6.35-803.8 upload inadvertantly emitted a linux-libc-dev
at this version.  This is far in the future compared to the master
branch leading to binary upload failures for armel when uploading
master.

Add a 2.6.35 and master branch specific bodge to bump the ABI number up
by 1000 just for the linux-libc-dev package.

Signed-off-by: Andy Whitcroft <apw@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>