ARM: vexpress: set RAM latencies to 1 cycle for PL310 on ct-ca9x4 tile
The PL310 on the ct-ca9x4 tile for the Versatile Express does not need
to add additional latency when accessing memory. Unfortunately, the
boot monitor sets this up for an 8-cycle delay on reads and writes,
resulting in restricted access to the bus for devices accessing memory
via uncached mappings. A symptom of this problem is display corruption
when the CLCD controller cannot read the framebuffer quickly enough
to display at a sensible refresh rate due to traffic from the cache
controller locking the bus.

This patch sets the L2 RAM latencies to the correct value of 1 cycle
on the ct-ca9x4 tile before enabling the L2 cache.

Cc: Russell King - ARM Linux <linux@arm.linux.org.uk>
Cc: Catalin Marinas <catalin.marinas@arm.com>
Signed-off-by: Will Deacon <will.deacon@arm.com>