UBUNTU: SAUCE: Fix fglrx compilation on 64bit after CVE-2010-3081
Patch the fglrx code to check for changes to compat_alloc_user_space and
adapt for it. Also adds code to follow kernel changes.
Based on the patch from Evan Broder but modified to do the testing in
the Makefile instead of make.sh as this is not used in L-R-M.

BugLink: http://bugs.launchpad.net/bugs/642518

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>