UBUNTU: [Config] Enable CONFIG_VMSPLIT_2G=y for OMAP4
BugLink: http://bugs.launchpad.net/bugs/633227

The default vm split is 3G:1G user:kernel, which will put 120M memory as
highmem in our OMAP4 Panda ES2.0. Highmem is not very stable currently,
we will meet the Bus Error bug when building kernel package on ES2.0.

This patch is a workaround, which let us to use full of 1G memory but
size of highmem is 0K with 2G:2G split. It's quite stable because
highmem is not active.

When we find the root cause and final fixing, we can switch back to
3G:1G normal split config.

Tested-by: Ricardo Salveti de Araujo <ricardo.salveti@canonical.com>
Signed-off-by: Bryan Wu <bryan.wu@canonical.com>