wlan: register wlan device only on pandaboard
BugLink: https://bugs.launchpad.net/bugs/665039

Registering the wlan device on blaze triggers the loading
of user-space wlan drivers for 1271 chip. These user-space
drivers are not compatible with the 1283 chip of the blaze and
cause a kernel crash.

The event is the same whatever the wlan controller, which
prevents managing different wlan drivers in user-space.

This patch registers the wlan device only on pandaboard.
Long term, we would like to generate different events depending
on the wlan controller.

Signed-off-by: Sebastien Jan <s-jan@ti.com>
Signed-off-by: Bryan Wu <bryan.wu@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>