SRU: Fix DVI for Beagle xM rev A3
SRU Justification:

Impact: With BeagleBoard xM rev A3 the DVI_PU is now used and connected to GPIO
2 on TPS65950. Without setting up this GPIO the DVI will be off, and the user
will be unable to install and use the distro.

Fix: The problem is fixed by setting up the GPIO 2 on TPS65950 for when running
at Beagle xM. There's no problem when setting this GPIO in previous xM revisions
as it's not connected. For more details please see
http://beagleboard.org/hardware/design.

Testcase: When booting with current kernel on xM A3 you should be unable to use
the DVI. After applying the fix it should work the same way as with previous xM revisions.

BugLink: https://bugs.launchpad.net/bugs/663642

Signed-off-by: Robert Nelson <robertcnelson@gmail.com>
Signed-off-by: Ricardo Salveti de Araujo <ricardo.salveti@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Acked-by: Leann Ogasawara <leann.ogasawara@canonical.com>