LINARO: Turn on and enforce PM CONFIG options
As requested by PM team turn on:
    CONFIG_PM
    CONFIG_REGULATOR
    CONFIG_PM_DEBUG
    CONFIG_DEBUG_FS
for all flavours.

Also add these to debian.linaro/config/enforce to ensure they
are always left on.

BugLink: https://bugs.launchpad.net/bugs/672791

Turn off CONFIG_SUSPEND for vexpress because that turns on
CONFIG_HOTSWAP_CPU which is not supported on vexpress.

Turn on CONFIG_NEON for vexpress

Signed-off-by: John Rigby <john.rigby@linaro.org>