UBUNTU: Add compat-wireless-2.6.37
BugLink: https://bugs.launchpad.net/bugs/673245

As part of the linux-backports-modules package, we periodically provide
an updated compat-wireless stack.  It's come to our attention that
future requests for newer versions of the compat-wireless stack will be
coming down the pipeline. In preparation, add infrastructure to start
tracking the 2.6.37 compat-wireless stack.  We'll wait to flip the bit
to actually build it until 2.6.37 is final.  Until then, expect periodic
updates as 2.6.37-rc#'s are released.

Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>