UBUNTU: Add compat-wireless-2.6.37
BugLink: http://bugs.launchpad.net/bugs/673245

As part of the linux-backports-modules package, we periodically provide
an updated compat-wireless stack. linux-backports-modules-2.6.32 already
has a pending pull request to add the 2.6.36 compat-wireless stack. It's
come to our attention that future requests for newer versions of the
compat-wireless stack will be coming down the pipeline. In preparation,
add infrastructure to start tracking the 2.6.37 compat-wireless stack.

Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>